requirements.txt
google-adk==1.14.0
google-cloud-logging==3.11.4
python-dotenv==1.0.1